requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
autopep8==2.2.0
binaryornot==0.4.4
certifi==2024.6.2
cffi==1.17.0
channels==4.1.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
cookiecutter==2.6.0
cryptography==43.0.0
daphne==4.1.2
distlib==0.3.8
Django==5.1
django-tenant-schemas==1.12.0
djangorestframework==3.15.2
djangorestframework-bulk==0.2.1
drf-yasg==1.21.7
filelock==3.14.0
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
inflection==0.5.1
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mysql-connector==2.2.9
mysql-connector-python==9.0.0
mysqlclient==2.2.4
ordered-set==4.1.0
packaging==24.1
pipenv==2023.12.1
platformdirs==4.2.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.11.1
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
service-identity==24.1.0
setuptools==74.0.0
six==1.16.0
sqlparse==0.5.1
text-unidecode==1.3
Twisted==24.7.0
txaio==23.1.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.1
virtualenv==20.26.2
zope.interface==7.0.3
